Add `User-Agent` to HTTP requests
It's a good practice to always send a customised user agent in order to
differentiate the traffic between sources. Adding a user agent with the
version number inside of it is also a valuable tool in order to assess
the versions of your tool still in use in the outside world.