Add boilerplate ajax call + placeholder for healthcheck content
This renders an empty healthcheck <div> on the "Help" tab. We then use jQuery to find when this <div> becomes visible (i.e. when the user navigates to the "Help" tab). Once the <div> becomes visible we populate it with some "Loading..." text and then fire off a nonced AJAX request. The HTML response from the AJAX request will be loaded into the <div> on success. Various errors are reported and logged to the console if the healthcheck fails completely. This should help users self diagnose any local issues.