Broken page error solution
Recently I see that the $theme['rating']  will return array if there is no ratings for a theme. So the code will show error and break the page to load. So I added the is_array() function. Also the return is float() if there are ratings so I added (int) converter to solve that issue too. This error appears in PHP v7.2 version.