docker-compose tweak: Install WP automatically locally
This will use the wpcli to install WordPress locally automatically.
We wait for the database to be healthy and then we run the "wp core install" command.

This will allow us to just run `docker-compose up` and then access a fully installed WordPress environment at http://localhost:8080

Default username and password are just dev/dev, but we could configure that via env variable overrides if needed.