docker-compose tweak: Mount entire plugin as volume
This mounts the current directory as a volume into the wp-content/plugins/envato-market/ folder.
Avoiding the need to install the WordPress plugin via zip method, the plugin will be available in WordPress straight away.

We could also extend the "wpcli" command to activate the plugin on first run as well, making the first "getting started" flow even nicer.